requirements.txt
django
djangorestframework
django-cors-headers
pyjwt
cryptography
requests